refactor: standardize converter module naming and extract shared test utility
Rename converter modules to follow consistent {format}.rs pattern:
- csv_conv.rs -> csv.rs
- json_conv.rs -> json.rs
- xml_conv.rs -> xml.rs

Extract duplicated normalize() function from 9 integration test files
into tests/common/mod.rs, eliminating ~135 lines of repeated code.
