+++ src/converter/mod.rs
@@ -1,14 +1,14 @@
-pub mod csv_conv;
+pub mod csv;
 pub mod docx;
 pub mod gemini;
 pub mod html;
 pub mod image;
-pub mod json_conv;
+pub mod json;
 pub(crate) mod ooxml_utils;
 pub mod plain_text;
 pub mod pptx;
 pub mod xlsx;
-pub mod xml_conv;
+pub mod xml;
 
 use std::sync::Arc;
